Agentic-RAG-Chatbot-for-Multi-Format-Document-QA-using-Model-Context-Protocol-main/requirements.txt
streamlit
sentence-transformers
faiss-cpu
pymupdf
python-pptx
python-docx
pandas
requests
